move diagram to lib folder
