+++ src/components/widgets/Node/NodeInPorts.tsx
@@ -6,7 +6,7 @@ import {
   PortWidget,
 } from '@projectstorm/react-diagrams';
 
-import { NodeModel } from '../../../diagram/models';
+import { NodeModel } from '../../../lib/diagram/models';
 import { Store } from '../../../store';
 
 interface Props {
