+++ src/components/widgets/Node/NodeOpenModalRequest.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useEffect } from 'react';
 import { observer } from 'mobx-react-lite';
-import { NodeModel } from '../../../diagram/models';
+import { NodeModel } from '../../../lib/diagram';
 import { Store, withStore } from '../../../store';
 
 interface Props {
