+++ src/components/widgets/Node/NodeWidgetInspectIcon.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 import { observer } from 'mobx-react-lite';
 
 import { Store } from '../../../store';
-import { NodeModel } from '../../../diagram/models';
+import { NodeModel } from '../../../lib/diagram';
 
 interface Props {
   store: Store;
