+++ src/lib/diagram/factories/index.ts
