+++ src/lib/diagram/index.ts
@@ -1,0 +1,2 @@
+export * from './factories';
+export * from './models';
