+++ src/lib/diagram/models/PortModel.ts
