+++ src/lib/diagram/models/index.ts
