+++ src/sections/App/AppHotkeys.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { observer } from 'mobx-react-lite';
 import { useHotkeys } from 'react-hotkeys-hook';
-import { NodeModel } from '../../diagram/models';
+import { NodeModel } from '../../lib/diagram';
 import { Store } from '../../store';
 
 interface Props {
