+++ src/sections/Toolbar/ToolbarInspectables.tsx
@@ -1,7 +1,7 @@
 import React, { FC, useState } from 'react';
 import { observer } from 'mobx-react-lite';
 import { Store } from '../../store';
-import { NodeModel } from '../../diagram/models';
+import { NodeModel } from '../../lib/diagram';
 
 interface Props {
   store: Store;
